requirements.txt
pyrogram==2.0.106
PyWinCtl==0.3
tgcrypto
requests
tabulate
pyromod==3.1.6
ffmpeg
mediainfo
cloudscraper
beautifulsoup4==4.12.2
bs4==0.0.2
aiohttp
aiofiles
aiohttp>=3.7.4
requests>=2.25.1
pyrogram>=1.2.9
pycryptodome>=3.10.1
cloudscraper>=1.2.58
PyJWT>=2.0.0
yt-dlp==2023.12.30
pycryptodome==3.20.0
PySocks==1.7.1
websockets==12.0
psutil
pymongo
async-lru==2.0.4
helper
pyaes==1.6.1
soupsieve==2.5
py-tgcalls==1.2.9
python-decouple==3.8
Flask==3.0.3
gunicorn==22.0.0
